Move browsing into a Browser class
