+++ src/Internal/Engine.php
@@ -7,7 +7,6 @@ namespace Loupe\Loupe\Internal;
 use Composer\InstalledVersions;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
-use Doctrine\DBAL\Query\QueryBuilder;
 use Loupe\Loupe\BrowseParameters;
 use Loupe\Loupe\BrowseResult;
 use Loupe\Loupe\Configuration;
@@ -20,6 +19,7 @@ use Loupe\Loupe\Internal\Index\Indexer;
 use Loupe\Loupe\Internal\Index\IndexInfo;
 use Loupe\Loupe\Internal\LanguageDetection\NitotmLanguageDetector;
 use Loupe\Loupe\Internal\LanguageDetection\PreselectedLanguageDetector;
+use Loupe\Loupe\Internal\Search\Browser;
 use Loupe\Loupe\Internal\Search\Searcher;
 use Loupe\Loupe\Internal\Search\Sorting\Relevance;
 use Loupe\Loupe\Internal\StateSetIndex\CachedStateSetIndex;
@@ -127,11 +127,7 @@ class Engine
         $this->maybeWrapStateSetIndexWithCache();
 
         try {
-            if ($this->canScanDocuments($parameters)) {
-                return $this->getConnection()->transactional(fn (): BrowseResult => $this->scanDocuments($parameters));
-            }
-
-            return (new Searcher($this, $this->filterParser, $parameters))->fetchResult();
+            return (new Browser($this, $this->filterParser, $parameters))->fetchResult();
         } catch (Exception $exception) {
             // If we need a re-index (e.g. schema has changed via an update from an old to a newer Loupe version)
             // we return an empty result. Otherwise, we want to see the exception
@@ -358,98 +354,6 @@ class Engine
             ->executeQuery('SELECT (SELECT page_count FROM pragma_page_count) * (SELECT page_size FROM pragma_page_size)')
             ->fetchOne()
         ;
-    }
-
-    private function scanDocuments(BrowseParameters $parameters): BrowseResult
-    {
-        $start = (int) floor(microtime(true) * 1000);
-
-        $limit = $parameters->getLimit();
-        $offset = $parameters->getOffset();
-
-        if (null !== $parameters->getHitsPerPage() || null !== $parameters->getPage()) {
-            $limit = $parameters->getHitsPerPage() ?? SearchParameters::MAX_LIMIT;
-            $offset = (($parameters->getPage() ?? 1) - 1) * $limit;
-        }
-
-        $documentsAlias = $this->indexInfo->getAliasForTable(IndexInfo::TABLE_NAME_DOCUMENTS);
-        $queryBuilder = $this->getConnection()->createQueryBuilder()
-            ->from(IndexInfo::TABLE_NAME_DOCUMENTS, $documentsAlias)
-            ->orderBy($documentsAlias.'._id', 'ASC')
-            ->setFirstResult($offset)
-            ->setMaxResults($limit)
-        ;
-
-        $hits = $this->fetchHits($queryBuilder, $documentsAlias, $parameters->getAttributesToRetrieve());
-        $totalHits = [] === $hits ? 0 : $this->countDocuments();
-        $totalPages = 0 === $limit ? 0 : (int) ceil($totalHits / $limit);
-        $currentPage = 0 === $limit ? 0 : (int) floor($offset / $limit) + 1;
-
-        return new BrowseResult(
-            $hits,
-            $parameters->getQuery(),
-            (int) floor(microtime(true) * 1000) - $start,
-            $limit,
-            $currentPage,
-            $totalPages,
-            $totalHits,
-        );
-    }
-
-    private function canScanDocuments(BrowseParameters $parameters): bool
-    {
-        return '' === $parameters->getQuery()
-            && '' === $parameters->getFilter()
-            && $parameters->getOffset() >= 0
-            && $parameters->getLimit() >= 0;
-    }
-
-    /**
-     * @param array<string> $attributesToRetrieve
-     *
-     * @return array<array<string, mixed>>
-     */
-    private function fetchHits(QueryBuilder $queryBuilder, string $documentsAlias, array $attributesToRetrieve): array
-    {
-        $primaryKey = $this->configuration->getPrimaryKey();
-
-        if ([$primaryKey] === $attributesToRetrieve) {
-            $ids = $queryBuilder
-                ->select($this->primaryKeyExpression($documentsAlias, $primaryKey))
-                ->setParameter('__loupe_pk_path', '$.'.$primaryKey)
-                ->fetchFirstColumn()
-            ;
-
-            return array_map(static fn (mixed $id): array => [$primaryKey => $id], $ids);
-        }
-
-        $documents = $queryBuilder->select($documentsAlias.'._document')->executeQuery()->iterateColumn();
-        $hits = [];
-
-        if (\in_array('*', $attributesToRetrieve, true)) {
-            foreach ($documents as $document) {
-                $hits[] = Util::decodeJson($document);
-            }
-
-            return $hits;
-        }
-
-        $keep = array_flip($attributesToRetrieve);
-
-        foreach ($documents as $document) {
-            $hits[] = array_intersect_key(Util::decodeJson($document), $keep);
-        }
-
-        return $hits;
-    }
-
-    private function primaryKeyExpression(string $documentsAlias, string $primaryKey): string
-    {
-        return match ($this->indexInfo->getDocumentSchema()[$primaryKey] ?? null) {
-            LoupeTypes::TYPE_STRING => $documentsAlias.'._user_id',
-            LoupeTypes::TYPE_NUMBER => \sprintf('CAST(%s._user_id AS NUMERIC)', $documentsAlias),
-            default => \sprintf('json_extract(%s._document, :__loupe_pk_path)', $documentsAlias),
-        };
     }
 
     private function executeIndexOperation(callable $operation): void
